electric_demand_forecast/requirements.txt
streamlit==1.32.2
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.0
xgboost>=2.0.3
matplotlib>=3.8.0
seaborn>=0.13.2
joblib>=1.3.2
shap>=0.45.0
scipy>=1.11.4
protobuf==3.20.3
pillow>=10.3.0